+++ source-code/shared/lib/server/router.ts
@@ -1,0 +1,13 @@
+import express from "express"
+import { openaiService } from "../openai/index.server.js"
+
+/**
+ * Entry point for all services that require a server.
+ *
+ * @example
+ *   const app = express()
+ *   app.use(inlangServices)
+ */
+export const router = express.Router()
+
+router.use(openaiService)